refactor: move getFilters() and getQuickFilters() api methods in folder filters
