+++ src/app/map/[category]/[subcategory]/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { getItems } from '@/app/map/[category]/page'
 import { CategoryPage, SubcategoryPage } from '@/types/Page'
-import { getQuickFilters } from '@/globals/api/methods/getFilters'
+import { getQuickFilters } from '@/globals/api'
 import ObjectsMap, { GetItemsForMapFn } from '@/app/map/_components/ObjectsMap'
 
 async function Page({ params }: SubcategoryPage & CategoryPage) {
