+++ src/globals/api/index.ts
@@ -1,11 +1,13 @@
 export * from './methods/catalog'
 export * from './methods/main-page'
 export * from './methods/catalog-details'
+export * from './methods/filters'
 export * from './methods/geo'
 
+export * from './methods/sendCallRequest'
+
+// Mocked
 export * from './methods/getAddresses'
 export * from './methods/getLinks'
 export * from './methods/getServices'
 export * from './methods/getSocials'
-export * from './methods/sendCallRequest'
-export * from './methods/getFilters'
