+++ src/globals/api/methods/filters/getQuickFilters.ts
@@ -1,13 +1,5 @@
 import { apiCall, APIResponse, SlashPath } from '@/globals/api/apiCall'
-import { Filters, QuickFilters } from '@/types/FiltersType'
-
-export async function getFilters(objectType: string) {
-  const path: SlashPath = `/filter/${objectType}`
-  const res = await apiCall<false, APIResponse<Filters>>(path, {
-    method: 'GET',
-  })
-  return res.data
-}
+import { QuickFilters } from '@/types/FiltersType'
 
 export async function getQuickFilters(objectType: string) {
   const path: SlashPath = `/filter/${objectType}/main`
