+++ src/globals/api/methods/filters/index.ts
@@ -1,0 +1,2 @@
+export * from './getFilters'
+export * from './getQuickFilters'
